chatbot/requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
et-xmlfile==1.1.0
filelock==3.0.12
Flask==2.0.1
future==0.18.2
fuzzychinese==0.1.5
gensim==3.8.1
huggingface-hub==0.0.17
idna==3.2
itsdangerous==2.0.1
jieba==0.42.1
Jinja2==3.0.1
joblib==1.0.1
line-bot-sdk==1.20.0
MarkupSafe==2.0.1
nltk==3.6.3
numpy==1.21.2
openpyxl==3.0.9
packaging==21.0
pandas==1.3.3
Pillow==8.3.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
regex==2021.8.28
requests==2.26.0
sacremoses==0.0.45
scikit-learn==1.0
scipy==1.7.1
sentencepiece==0.1.96
six==1.16.0
smart-open==5.2.1
threadpoolctl==3.0.0
torch==1.9.0
torchvision==0.10.0
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.6
Werkzeug==2.0.1
